Add sequential action
A sequential action follows a composite pattern allowing for a richer
use of actions.

One of the problems here is that MongoDB does not correctly serialize
the primitive.A to a []any. This causes a leakage of primitives into the
model... This is currently stopped by manually copying it into a slice
in the model which is pretty dirty to me.