Do not throw an error if a key does not exists
This is currently just done by swallowing the error.

Another option would be to use the option field accessor. The problem
here is that this adds a lot of complexity to the rules as iterating
over the options is not possible. This should be paired with an exists,
making it a lot more complex.