Register plugin in action factory
This allows the use of plugins in Reflexes. This still feels a bit
cluncky to use and needs to be refined along the way. The logic of
generating a plugin action should be extracted so it is reusable.

Next, the plugins lifecylce needs to be checked as well. Otherwise the
plugins will persist even when the server terminated.