Read TheHive config from environment
Two configs are needed for TheHive plugin. First an action config is
needed which is supplied through the security ctx. Secondly, an TheHive
config is needed for information concerning an TheHive instance. As
these contain secrets like API keys, they are read from the env instead
of the DB.

This commit also split TheHive plugin over multiple go files which
required an alteration to the makefile.